feat: add icons

+++ api/controller/postController.js
@@ -87,7 +87,18 @@ const editPost = async (req, res) => {
 };
 const deletePost = async (req, res) => {
   const { id } = req.params;
+  const tokenData = await jwt.verify(
+    req.cookies["access-token"],
+    process.env.ACCESS_TOKEN_SECRET
+  );
   try {
+    const dbResult = await Post.findById(id);
+    const isAuthor =
+      JSON.stringify(dbResult.author) == JSON.stringify(tokenData.id);
+
+    if (!isAuthor) {
+      return res.status(400).json("You are not author");
+    }
     const result = await Post.findByIdAndDelete(id);
     return res.status(200).json(result);
   } catch (err) {
